% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHMM.R
\name{AHMM}
\alias{AHMM}
\title{Treatment effect estimation for AHMM with 2-level clustering}
\usage{
AHMM(
  formula,
  cluster,
  data,
  subset,
  method = c("z", "t"),
  bcv = c("CZ", "KC", "MD", "FG", "MBN"),
  na.action
)
}
\arguments{
\item{formula}{Model specification in the traditional 'Y ~ X' sense,
where Y is in Surv format (function in 'survival' package).}

\item{cluster}{Variable in the data that contains information about group
(cluster) assignment of each individual observation}

\item{data}{Input data source}

\item{subset}{If only a subset of the data shall be used}

\item{method}{Test that should be used in the estimation procedure. Input
"z" if z-test or "t" if t-test should be used respectively.}

\item{bcv}{Input specifying which variance correction should be used in
the estimation procedure. "CZ" is the uncorrected variance as in Cai and
Zeng, "KC", "MD", "FG", and "MBN" are variance corrections according to
Kauermann and Carroll, Mancl and DeRouen, Fay and Graubard, Morel et al.
respectively.}

\item{na.action}{Specify how missing values should be handled.}
}
\value{
This function treatment effect estimate of the underlying AHMM.
}
\description{
This function conducts an estimation procedure of treatment effect
in an Additive Hazards Mixed Model with two-level clustering
and no additional covariates. The underlying methods are described in
(reference our paper).
}
\details{
Lots of descriptive details could go here. This line just holds the place.
}
\examples{
AHMM(Surv(U,Delta)~Z, cluster = ID, data = Test_Data)



}
\keyword{methods}
