% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermTest.R
\name{PermTest}
\alias{PermTest}
\title{Permutation test calculation for AHMM with 2-level clustering}
\usage{
PermTest(ID, Z_cl, U, Delta, pmt = NULL)
}
\arguments{
\item{ID}{Vector specifying in which group/cluster each individual
sample belongs to}

\item{Z_cl}{Vector of actual allocation into treatment arms}

\item{U}{Vector with observed event times}

\item{Delta}{Censoring vector}

\item{pmt}{User specified permutation matrix}
}
\value{
This function returns returns randomization-based test p-value
of the underlying AHMM
}
\description{
This function performs a permutation test calculation for an underlying test of
treatment effect under the Additive Hazards Mixed Model with two-level
clustering and randomization occurring at the higher (cluster) level.
The underlying methods are described in (2022 Blaha, Esserman, Li paper).
}
\examples{
PermTest(ID=Test_Data$ID, Z_cl=trt, U=Test_Data$U, Delta=Test_Data$Delta, pmt = pmt)



}
\keyword{methods}
