% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSize.R
\name{SampleSize}
\alias{SampleSize}
\title{A Sample size calculation for AHMM with 2-level clustering and no covariates}
\usage{
SampleSize(
  lambda0,
  delta,
  sigma2,
  tau,
  m_bar,
  CV = 0,
  pi = 0.5,
  alpha = 0.05,
  beta = 0.2,
  test = c("z", "t")
)
}
\arguments{
\item{lambda0}{Expected baseline hazard (scalar)}

\item{delta}{Expected treatment effect on hazard difference scale (scalar)}

\item{sigma2}{Expected clustering effect - variance of the frailty parameter}

\item{tau}{Planned duration of the study}

\item{m_bar}{Expected average cluster size}

\item{CV}{Expected coefficient of variation among cluster sizes. Default
is 0.}

\item{pi}{Randomization ratio (arm B vs. A). Accepted values are between 0
and 1. Larger values than 0.5 indicated that more subjects are allocated
into the experimental arm. Smaller values than 0.5 indicated that more
subjects are allocated into the control arm. Default is 0.5.}

\item{alpha}{Desired size of the underlying test. Default is alpha=0.05.}

\item{beta}{Desired power of the test. Default is 80\%.}

\item{test}{Choice between "z" for large sample z-test and "t" for small
sample t-test. Default is c("t","z").}
}
\value{
This function returns returns required sample size of the underlying
test of hypothesis H_0:delta=0 vs. two-sided alternative for either t-test
whe test="t" is selected or z-test if test="z" is selected, given the
desired power and size of the test.
}
\description{
This function performs a sample size calculation for a Additive Hazards
Mixed Model with two-level clustering and no additional covariates.
The underlying methods are described in (Blaha et al. paper).
}
\details{
Lots of descriptive details could go here. This line just holds the place.
}
\examples{
SampleSize(lambda0=4,delta=1.5,sigma2=0.5,tau=1,m_bar=10)



}
\keyword{methods}
