% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIF.R
\name{VIF}
\alias{VIF}
\title{VIF and ICC calculation for AHMM with 2-level clustering}
\usage{
VIF(
  lambda0,
  delta,
  sigma2,
  tau,
  m_bar,
  CV = 0,
  pi = 0.5,
  alpha = 0.05,
  beta = 0.2
)
}
\arguments{
\item{lambda0}{Anticipated baseline hazard (scalar)}

\item{delta}{Treatment effect on hazard difference scale (scalar)}

\item{sigma2}{Clustering effect - variance of the frailty parameter}

\item{tau}{Duration of the planned study (administrative censoring)}

\item{m_bar}{Anticipated average cluster size}

\item{CV}{Coefficient of variation among cluster sizes. Default is 0.}

\item{pi}{Randomization ratio (arm B vs. A). Accepted values are between 0
and 1. Larger values than 0.5 indicated that more subjects are allocated
into the experimental arm. Smaller values than 0.5 indicated that more
subjects are allocated into the control arm. Default is 0.5.}

\item{alpha}{Desired size of the underlying test. Default is alpha=0.05.}

\item{beta}{Desired power of the test. Default is 80\%.}
}
\value{
This function returns returns VIF and ICC of the underlying AHMM.
}
\description{
This function performs a VIF and ICC calculation for an underlying test of
treatment effect under the Additive Hazards Mixed Model with two-level
clustering and randomization occurring at the higher (cluster) level.
The underlying methods are described in (2022 Blaha, Esserman, Li paper).
}
\examples{
VIF(lambda0=4,delta=1.5,sigma2=0.5,tau=1,m_bar=10,CV=0.25)



}
\keyword{methods}
