% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSize.R
\name{SampleSize}
\alias{SampleSize}
\title{A Sample size calculation for AHMM with 2-level clustering}
\usage{
SampleSize(
  lambda0,
  delta,
  sigma2,
  tau,
  m_bar,
  CV = 0,
  pi = 0.5,
  alpha = 0.05,
  beta = 0.2,
  test = c("z", "t")
)
}
\arguments{
\item{lambda0}{Anticipated baseline hazard (scalar)}

\item{delta}{Treatment effect on hazard difference scale (scalar)}

\item{sigma2}{Clustering effect - variance of the frailty parameter}

\item{tau}{Duration of the planned study (administrative censoring)}

\item{m_bar}{Anticipated average cluster size}

\item{CV}{Coefficient of variation among cluster sizes. Default is 0.}

\item{pi}{Randomization ratio (arm B vs. A). Accepted values are between 0
and 1. Larger values than 0.5 indicated that more subjects are allocated
into the experimental arm. Smaller values than 0.5 indicated that more
subjects are allocated into the control arm. Default is 0.5.}

\item{alpha}{Desired size of the underlying test. Default is alpha=0.05.}

\item{beta}{Desired power of the test. Default is 80\%.}

\item{test}{Choice between "z" for large sample z-test and "t" for small
sample t-test. Default is c("t","z").}
}
\value{
This function returns required sample size of the underlying
test of hypothesis H_0:delta=0 vs. two-sided alternative for either t-test
when test="t" is selected or z-test if test="z" is selected, given the
desired power and size of the test.
}
\description{
This function performs sample size calculation for an underlying test of
treatment effect under the Additive Hazards Mixed Model with two-level
clustering and randomization occurring at the higher (cluster) level.
The underlying methods are described in (2022 Blaha, Esserman, Li paper).
}
\examples{
SampleSize(lambda0=4,delta=1.5,sigma2=0.5,tau=1,m_bar=10)



}
\keyword{methods}
